client/requirements.txt
scapy
requests
python-socketio
psutil